Introduce a settings storage
This provides a place to store user-specific settings for LibMan.  These settings will live under the user's home folder and apply to all projects on that machine.

Along the way, I set a version for Moq (missed it before so NuGet resolved to the oldest available) and bumped up the FxCopAnalyzers to address a false positive.