Fix tests to account for prior changes in NPM package search.
In an earlier PR, the search service was changed to use the NPM registry.  Because this returns results based on fulltext matches in the registry, it often returns package names that don't contain the search text (because their metadata matches the search term instead).  We expanded the size of the result set to 100, so that it would include more names that match the search text.

Also, the new search service doesn't return any results for an empty search string.