Support file glob patterns in "files" property
This takes the initial ILibraryInstallationState's Files, and expand any patterns into the file names that they match.  It returns a new LibraryInstallationState instance, which required removing some AreSame checks in the tests.

If a pattern matches 0 files, it will simply be removed from the set of files.  All non-pattern file names are preserved, and will continue to be checked as valid file names.