Workaround an assembly version conflict (#533)
Moq had a dependency on System.Threading.Tasks.Extensions 4.2.0.0, but our dependency graph pull in 4.2.0.1 (thanks, whoever bumped that).  Setting the binding redirect in the test project fixes it at runtime.

This was likely a regression caused in c5beab72, when we bumped up the version of Microsoft.VisualStudio.Threading (which seems to be at least one source of the 4.2.0.1 transitive dependency).