Make CopyFileAsync actually async.
Also made the HostInteractions.CopyFileAsync implementations more consistent, including adding clarification that the destinationPath parameter is relative to the WorkingDirectory (where libman.json is located).  Previously we would have written the files relative to the process working directory. In the CLI this is fine, as they're the same; in MSBuild and VS, these might not be the case.