Used stale cached Cdnjs metadata if request fails
Currently we download these files and cache them locally with a hardcoded lifetime.  Once the file is expired, we attempt to download a new copy; however, if that request fails, we just fail the operation.  With this change, we'll at least fall back to the already-stored copy, which may be stale but should still work in most cases.