Move localization process to OneLocBuild (#616)
* Add new build step for OneLocBuild process

Because this build step modifies the resource files in-situ, it breaks
build reproducibility.  Moving it to the end should still produce the
outputs necessary to create a PR for localization handbacks.

This required renaming one resx file so that the new tool would generate the correct localized file names.

* Remove MAT-related XLF files

These will be replaced with LCL files from OneLocBuild soon.

We still have XLF files for the VSCT file, which uses a different tool
for conversion.