Rename source.extension.resx back to have the "en" language ID
Removing this in e2cb34ae28a37 caused the VSCT command items to no
longer show up in the IDE.  I'm not sure why the file name matters here
(especially since documentation examples just use it without the
language ID, but maybe it has to do with supporting multiple
languages?).

As I recall, the rename happened in order to keep the localization tools
happy, using the LangIdOnName CopyOption specified in
localize/LocProject.json.  So while renaming it back, I've also adjusted
the build to copy the file to the name needed for the localization tools
to run; this is just like the .vsct file itself.